requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.2
pyttsx3==2.90
SpeechRecognition==3.10.0
pyaudio==0.2.13 